backend/Tracker/src/main/java/BudgetTracker/Tracker/service/UserService.java
package BudgetTracker.Tracker.service;

import BudgetTracker.Tracker.entity.User;
import BudgetTracker.Tracker.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class UserService {
    private UserRepository userRepository;

    @Autowired
    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public List<User> getAllUsers() {
        return userRepository.findAll();
    }

    public User findUserById(Long Id) { return userRepository.findById(Id).orElse(null);}

    public User createNewUser(User user) {
        return userRepository.save(user);
    }

    public Optional<User> findUserByNameAndEmail(String name, String email) {
        return userRepository.findByNameAndEmail(name, email);
    }

}
